IOS-3984: Restoring all sessions after login via card and via biometry
Removing and disconnecting all sessions for deleted card from repo
