+++ Tangem/App/Services/UserWalletRepository/CommonUserWalletRepository.swift
@@ -13,7 +13,7 @@ import TangemSdk
 
 class CommonUserWalletRepository: UserWalletRepository {
     @Injected(\.tangemApiService) private var tangemApiService: TangemApiService
-    @Injected(\.walletConnectService) private var walletConnectServiceProvider: WalletConnectService
+    @Injected(\.walletConnectService) private var walletConnectService: WalletConnectService
     @Injected(\.failedScanTracker) var failedCardScanTracker: FailedScanTrackable
     @Injected(\.analyticsContext) var analyticsContext: AnalyticsContext
 
@@ -342,7 +342,7 @@ class CommonUserWalletRepository: UserWalletRepository {
             }
         }
 
-        walletConnectServiceProvider.disconnectAllSessionsForUserWallet(with: userWalletId.toHexString())
+        walletConnectService.disconnectAllSessionsForUserWallet(with: userWalletId.toHexString())
         sendEvent(.deleted(userWalletId: userWalletId))
     }
 
@@ -412,6 +412,7 @@ class CommonUserWalletRepository: UserWalletRepository {
                     self.userWallets = self.savedUserWallets(withSensitiveData: true)
                     self.loadModels()
                     self.initializeServicesForSelectedModel()
+                    self.walletConnectService.initialize()
                     self.selectedModel?.initialUpdate()
 
                     if let selectedModel = self.selectedModel {
@@ -487,6 +488,7 @@ class CommonUserWalletRepository: UserWalletRepository {
 
                 setSelectedUserWalletId(savedUserWallet.userWalletId, reason: .userSelected)
                 initializeServicesForSelectedModel()
+                walletConnectService.initialize()
                 selectedModel?.initialUpdate()
 
                 sendEvent(.updated(userWalletModel: cardModel))
