+++ Tangem/App/Services/WalletConnect/WalletConnectService.swift
@@ -9,75 +9,23 @@
 import Foundation
 import Combine
 
-class CommonWalletConnectService {
-    @Injected(\.incomingActionManager) private var incomingActionManager: IncomingActionManaging
+protocol WalletConnectService {
+    var canEstablishNewSessionPublisher: AnyPublisher<Bool, Never> { get }
+    var newSessions: AsyncStream<[WalletConnectSavedSession]> { get async }
 
-    private var v2Service: WalletConnectV2Service
-
-    init() {
-        v2Service = WalletConnectFactory().createWCService()
-    }
+    func initialize()
+    func openSession(with uri: WalletConnectRequestURI)
+    func disconnectSession(with id: Int) async
+    func disconnectAllSessionsForUserWallet(with userWalletId: String)
 }
 
-extension CommonWalletConnectService: WalletConnectService {
-    var canEstablishNewSessionPublisher: AnyPublisher<Bool, Never> {
-        v2Service.canEstablishNewSessionPublisher.eraseToAnyPublisher()
-            .eraseToAnyPublisher()
-    }
-
-    var newSessions: AsyncStream<[WalletConnectSavedSession]> {
-        get async {
-            await v2Service.newSessions
-        }
-    }
-
-//
-//    func initialize(with cardModel: CardViewModel) {
-//        guard cardModel.shouldShowWC else {
-//            return
-//        }
-//
-//        // Note: If we are planning to write unit tests for each class this factory can be wrapped
-//        // with protocol and injected via initializer. But for now I think it'll be enough.
-//        v2Service = WalletConnectFactory().createWCService(for: cardModel)
-//        incomingActionManager.becomeFirstResponder(self)
-//    }
-
-//    func reset() {
-//        incomingActionManager.resignFirstResponder(self)
-//        v2Service = nil
-//    }
-
-    func openSession(with uri: WalletConnectRequestURI) {
-        switch uri {
-        case .v2(let v2URI):
-            v2Service.openSession(with: v2URI)
-        }
-    }
-
-    func disconnectSession(with id: Int) async {
-        await v2Service.disconnectSession(with: id)
-    }
-
-    func disconnectAllSessionsForUserWallet(with userWalletId: String) {}
-
-//    func canOpenSession(with uri: WalletConnectRequestURI) -> Bool {
-//        switch uri {
-//        case .v2:
-//            return v2Service.canEstablishNewSessionPublisher.
-//        }
-//    }
+private struct WalletConnectServicingKey: InjectionKey {
+    static var currentValue: WalletConnectService = CommonWalletConnectService()
 }
 
-// MARK: - IncomingActionResponder
-
-extension CommonWalletConnectService: IncomingActionResponder {
-    func didReceiveIncomingAction(_ action: IncomingAction) -> Bool {
-        guard case .walletConnect(let uri) = action else {
-            return false
-        }
-
-        openSession(with: uri)
-        return true
+extension InjectedValues {
+    var walletConnectService: WalletConnectService {
+        get { Self[WalletConnectServicingKey.self] }
+        set { Self[WalletConnectServicingKey.self] = newValue }
     }
 }
