+++ Tangem/App/Services/WalletConnect/WalletConnectV2Service/Models/WalletConnectV2Error.swift
@@ -26,6 +26,8 @@ enum WalletConnectV2Error: LocalizedError {
     case unsupportedDApp
     case missingActiveUserWalletModel
     case missingUserWalletWithTargetId
+    case userWalletRepositoryIsLocked
+    case userWalletIsLocked
 
     case unknown(String)
 
@@ -48,6 +50,8 @@ enum WalletConnectV2Error: LocalizedError {
         case .unsupportedDApp: return 8015
         case .missingActiveUserWalletModel: return 8016
         case .missingUserWalletWithTargetId: return 8017
+        case .userWalletRepositoryIsLocked: return 8018
+        case .userWalletIsLocked: return 8019
 
         case .unknown: return 8999
         }
